# Backend/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "University Evaluation API",
    "version": "1.0.0",
    "description": "University evaluation system with 3 dimensions: Governance, Social, Environmental - Auto-generated documentation",
    "contact": {
      "name": "API Support",
      "email": "support@api.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "LOGIN_ERROR"
          },
          "message": {
            "type": "string",
            "example": "Invalid credentials"
          }
        }
      },
...
      "Dimension": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Governance"
          },
          "code": {
            "type": "string",
            "example": "governance"
          },
          "description": {
            "type": "string",
            "example": "Governance dimension evaluation"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-01T10:00:00Z"
          }
        }
      },
      "Question": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "dimension_id": {
            "type": "integer",
            "example": 1
          },
          "text": {
            "type": "string",
            "example": "How would you rate the university governance transparency?"
          },
          "order_index": {
            "type": "integer",
            "example": 1
          },
          "scale_descriptions": {
            "type": "object",
            "example": {
              "1": "Poor transparency",
              "2": "Fair transparency",
              "3": "Good transparency",
              "4": "Very good transparency",
              "5": "Excellent transparency"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-01T10:00:00Z"
          }
        }
      },
      "University": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Universidad Nacional"
          },
          "city": {
            "type": "string",
            "example": "Bogotá"
          },
          "department": {
            "type": "string",
            "example": "Cundinamarca"
          }
        }
      },
      "EvaluationResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "evaluation_id": {
            "type": "integer",
            "example": 1
          },
          "question_id": {
            "type": "integer",
            "example": 1
          },
          "score": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "example": 4
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-01T10:00:00Z"
          },
          "questions": {
            "$ref": "#/components/schemas/Question"
          }
        }
      },
      "Evaluation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "university_id": {
            "type": "integer",
            "example": 1
          },
          "dimension_id": {
            "type": "integer",
            "example": 1
          },
          "comments": {
            "type": "string",
            "nullable": true,
            "example": "Overall good governance practices"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-01T09:00:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-01T09:30:00Z"
          },
          "universities": {
            "$ref": "#/components/schemas/University"
          },
          "dimensions": {
            "$ref": "#/components/schemas/Dimension"
          }
        }
      },
      "EvaluationWithResponses": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Evaluation"
          },
          {
            "type": "object",
            "properties": {
              "evaluation_responses": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            }
          }
        ]
      }
    }
  },
  "paths": {
...
    "/evaluations": {
      "get": {
        "tags": [
          "Evaluations"
        ],
        "summary": "Get user's evaluations",
        "description": "Retrieve all evaluations created by the authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Evaluations retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Evaluations retrieved successfully"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "example": 1
                          },
                          "user_id": {
                            "type": "string",
                            "format": "uuid",
                            "example": "123e4567-e89b-12d3-a456-426614174000"
                          },
                          "university_id": {
                            "type": "integer",
                            "example": 1
                          },
                          "dimension_id": {
                            "type": "integer",
                            "example": 1
                          },
                          "comments": {
                            "type": "string",
                            "example": "Overall good governance practices"
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2024-01-01T09:00:00Z"
                          },
                          "updated_at": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2024-01-01T09:30:00Z"
                          },
                          "universities": {
                            "type": "object",
                            "properties": {
                              "name": {
                                "type": "string",
                                "example": "Universidad Nacional"
                              },
                              "city": {
                                "type": "string",
                                "example": "Bogotá"
                              },
                              "department": {
                                "type": "string",
                                "example": "Cundinamarca"
                              }
                            }
                          },
                          "dimensions": {
                            "type": "object",
                            "properties": {
                              "name": {
                                "type": "string",
                                "example": "Governance"
                              },
                              "code": {
                                "type": "string",
                                "example": "governance"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Evaluations"
        ],
        "summary": "Create or update evaluation",
        "description": "Create a new evaluation or update an existing evaluation. Each user can only have one evaluation per university per dimension.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "university_id",
                  "dimension_id",
                  "responses"
                ],
                "properties": {
                  "university_id": {
                    "type": "integer",
                    "description": "ID of the university being evaluated",
                    "example": 1
                  },
                  "dimension_id": {
                    "type": "integer",
                    "description": "ID of the dimension being evaluated",
                    "example": 1
                  },
                  "responses": {
                    "type": "array",
                    "description": "Array of question responses",
                    "items": {
                      "type": "object",
                      "required": [
                        "question_id",
                        "score"
                      ],
                      "properties": {
                        "question_id": {
                          "type": "integer",
                          "description": "ID of the question being answered",
                          "example": 1
                        },
                        "score": {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 5,
                          "description": "Score given to the question (1-5)",
                          "example": 4
                        }
                      }
                    }
                  },
                  "comments": {
                    "type": "string",
                    "description": "Optional overall comments about the evaluation",
                    "example": "The university shows strong governance practices with room for improvement in transparency"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Evaluation updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Evaluation updated successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/EvaluationWithResponses"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": "Evaluation created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Evaluation created successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/EvaluationWithResponses"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/evaluations/{id}": {
      "get": {
        "tags": [
          "Evaluations"
        ],
        "summary": "Get specific evaluation with responses",
        "description": "Retrieve a specific evaluation including all question responses",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the evaluation to retrieve",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Evaluation retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Evaluation retrieved successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/EvaluationWithResponses"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Evaluation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Evaluations"
        ],
        "summary": "Delete evaluation",
        "description": "Delete an evaluation.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the evaluation to delete",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Evaluation deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Evaluation deleted successfully"
                    },
                    "data": {
                      "type": null,
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Evaluation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}